Corregir las clases repository

+++ src/main/java/com/translogistics/parcial/repository/PlanillaViajeRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlanillaViajeRepository extends JpaRepository<PlanillaViaje, Integer>{
-
+public interface PlanillaViajeRepository extends JpaRepository<PlanillaViaje, Long>{
 }
